refacrtor: change to mamdani service WIP
