+++ src/app/routes/fuzzy/fuzzy.component.html
@@ -1,94 +1,16 @@
 <div class="container">
   <app-project-name [projectName]="projectName"></app-project-name>
   <section>
-    <form [formGroup]="variablesForm" (ngSubmit)="addVariable($event)">
-      <mat-card class="fuzzy-card">
-        <mat-card-content>
-          <mat-radio-group
-            aria-label="Select an option"
-            [(ngModel)]="variables.variableType"
-            [ngModelOptions]="{ standalone: true }"
-          >
-            <mat-radio-button value="input">Input</mat-radio-button>
-            <mat-radio-button value="output">Output</mat-radio-button>
-          </mat-radio-group>
-          <br />
-          <div class="row">
-            <div class="col-6">
-              <mat-form-field class="input-full-width">
-                <mat-label>Variable Name</mat-label>
-                <input matInput placeholder="Name" formControlName="name" />
-                <mat-error
-                  *ngIf="variablesForm.get('name').hasError('required')"
-                >
-                  Name is required
-                </mat-error>
-              </mat-form-field>
-            </div>
-            <div class="col-6">
-              <mat-form-field class="input-full-width">
-                <mat-label>Number of fuzzyAreas</mat-label>
-                <input
-                  matInput
-                  placeholder="fuzzy areas count"
-                  type="number"
-                  min="0"
-                  formControlName="fuzzyAreasCount"
-                />
-              </mat-form-field>
-            </div>
-          </div>
-          <br />
-          <div class="row">
-            <div class="col-6">
-              <mat-form-field class="input-full-width">
-                <mat-label>Range: Start</mat-label>
-                <input
-                  matInput
-                  placeholder="Start"
-                  type="number"
-                  min="0"
-                  formControlName="start"
-                />
-              </mat-form-field>
-            </div>
-            <div class="col-6">
-              <mat-form-field class="input-full-width">
-                <mat-label>Range: End</mat-label>
-                <input
-                  matInput
-                  placeholder="End"
-                  type="number"
-                  min="0"
-                  formControlName="end"
-                />
-              </mat-form-field>
-            </div>
-          </div>
-        </mat-card-content>
-        <mat-card-actions>
-          <button
-            mat-raised-button
-            color="primary"
-            type="submit"
-            [disabled]="!variablesForm.valid"
-          >
-            Add variable
-          </button>
-        </mat-card-actions>
-      </mat-card>
-    </form>
+    <app-variables-form></app-variables-form>
 
     <app-variables-list
       [title]="'Input Variables'"
-      [variables]="variables.inputs"
-      (addFuzzyArea)="addFuzzyArea($event)"
+      [variables]="mamdaniService.variables.inputs"
     ></app-variables-list>
 
     <app-variables-list
       [title]="'Output Variables'"
-      [variables]="variables.outputs"
-      (addFuzzyArea)="addFuzzyArea($event)"
+      [variables]="mamdaniService.variables.outputs"
     ></app-variables-list>
   </section>
 
@@ -97,7 +19,7 @@
       <mat-card class="fuzzy-card">
         <mat-card-title>Provide example for your variables:</mat-card-title>
         <mat-card-content>
-          <div *ngFor="let input of variables.inputs">
+          <div *ngFor="let input of mamdaniService.variables.inputs">
             <mat-form-field class="input-full-width">
               <mat-label>Name</mat-label>
               <input
@@ -120,7 +42,7 @@
     </form>
   </section>
 
-  <section id="rules" *ngIf="variables.inputs.length">
+  <section id="rules" *ngIf="mamdaniService.variables.inputs.length">
     <mat-card class="fuzzy-card">
       <mat-card-content>
         <form (ngSubmit)="createRule()" #form="ngForm">
@@ -138,7 +60,12 @@
           <div class="row">
             <div class="col-1"><p style="margin-top: 20px">IF</p></div>
             <div class="col-7">
-              <span *ngFor="let input of variables.inputs; let index = index">
+              <span
+                *ngFor="
+                  let input of mamdaniService.variables.inputs;
+                  let index = index
+                "
+              >
                 <span
                   style="margin: 0 10px"
                   *ngIf="index > 0"
@@ -164,7 +91,10 @@
             <div class="col-3">
               <mat-form-field appearance="fill">
                 <mat-select
-                  *ngFor="let output of variables.outputs; let index = index"
+                  *ngFor="
+                    let output of mamdaniService.variables.outputs;
+                    let index = index
+                  "
                   [(ngModel)]="rules.newRule.fuzzyAreas.output[index]"
                   [ngModelOptions]="{ standalone: true }"
                 >
