--- src/app/routes/fuzzy/fuzzy.component.scss
@@ -1,3 +1,0 @@
-.mat-radio-button ~ .mat-radio-button {
-  margin-left: 16px;
-}
