+++ src/app/routes/fuzzy/variables-form/variables-form.component.scss
@@ -1,0 +1,3 @@
+.mat-radio-button ~ .mat-radio-button {
+  margin-left: 16px;
+}
